fresh fork

+++ mt32-pi/external/circle-stdlib/libs/circle-newlib/winsup/testsuite/winsup.api/pthread/mutex4.c
@@ -17,7 +17,7 @@ static pthread_mutex_t mutex1;
  
 void * unlocker(void * arg)
 {
-  int expectedResult = (int)(size_t)arg;
+  int expectedResult = (int) arg;
 
   wasHere++;
   assert(pthread_mutex_unlock(&mutex1) == expectedResult);
